+++ migrations/0001_initial.sql
@@ -20,3 +20,8 @@ CREATE TABLE auth_user (
     password      TEXT    NOT NULL DEFAULT '',
     first_name    TEXT    NOT NULL DEFAULT '',
     last_name     TEXT    NOT NULL DEFAULT '',
+    is_staff      INTEGER NOT NULL DEFAULT 0,
+    is_active     INTEGER NOT NULL DEFAULT 1,
+    is_superuser  INTEGER NOT NULL DEFAULT 0,
+    last_login    TEXT,
+    date_joined   TEXT    NOT NULL